requirements.txt
langchain_chroma==0.2.4
langchain_community==0.3.26
langchain_huggingface==0.3.0
langchain_text_splitters==0.3.8
pymupdf4llm==0.0.25
python-dotenv==1.1.0
Requests==2.32.4
